+++ frontend/src/img/brevity_banner.png
