+++ frontend/src/img/brevity_icon.png
