+++ frontend/src/img/brevity_icon_dark.png
