+++ frontend/src/img/brevity_logo.png
